Refactored the codebase so now running inputs through a model and saving activations is a method of ModelHistory.
